rss_parser/episode_processor/requirements.txt
faster-whisper>=1.0.0
flask>=3.0
gunicorn>=22.0
google-cloud-storage>=2.0
requests>=2.31
